requirements.txt
datasets
transformers
torchtune
torchao

# Megatron
megatron-core
# tensorstore
# zarr